Changed parse_data so it used thread_last
